updated accounts migration

+++ db/migrations/20230405191444-create_accounts_table.sql
@@ -5,10 +5,12 @@ CREATE TABLE IF NOT EXISTS accounts
     id        serial primary key NOT NULL,
     first_name            varchar(50) NOT NULL,
     last_name            varchar(50) NOT NULL,
-    number serial,
-    balance serial,
+    number numeric,
+    balance numeric,
     created_at timestamp
 );
 
 -- +migrate Down
 DROP TABLE IF EXISTS accounts;
+
+
